refactor: move providers/ into features/completion/
- Delete src/providers/ and src/util/includeAutoCompletion.ts
- Move completion providers into src/features/completion/ for consistency
  with the rest of features/ (documentSymbolProvider, foldingProvider, etc.)
- Rename files to camelCase (targetPathCompletionProvider, bibtexCompletionProvider, xrefCompletionProvider)
- Rename AsciidocTargetPathAutoCompletionMonitor → AsciidocCompletionProviders
- Remove dead EventEmitter from the registration class
- Fix DocumentSelector: completion providers now use asciidocDocumentSelector
  (file + untitled) so completions work on unsaved documents
