+++ src/features/completion/bibtexCompletionProvider.ts
@@ -1,7 +1,7 @@
 import bibtexParse from '@orcid/bibtex-parse-js'
 import { readFileSync } from 'fs'
 import * as vscode from 'vscode'
-import { findFiles } from '../util/findFiles.js'
+import { findFiles } from '../../util/findFiles.js'
 import { Context, createContext } from './createContext.js'
 
 export const BibtexProvider = {
